Move auth.js out of plugins folder
