+++ auth.js
@@ -1,9 +1,9 @@
 
 const passport = require('passport');
-const helpers = require('../shared/helpers');
-const auth0 = require('../lib/services/auth0');
+const helpers = require('./shared/helpers');
+const auth0 = require('./lib/services/auth0');
 const Auth = auth0.Auth;
-const users = require('../controllers/users');
+const users = require('./controllers/users');
 const {check, validationResult, body} = require('express-validator/check');
 
 module.exports = {
