+++ lib/routes.js
@@ -13,7 +13,7 @@ const config = require('./config');
 const indexing = require('../controllers/indexing');
 const users = require('../controllers/users');
 const keystonePlugin = require('../plugins/keystone');
-const authPlugin = require('../plugins/auth');
+const authPlugin = require('../auth');
 
 /**
  * Application routes
